+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 - feat: Added l1->l2 messaging
+- test: add unitests primitives
 - tests: add tests for the rpcs endpoints
 - fix: pending contract storage not stored properly
 - test: add tests crate `db`